master - files for lesson 2 and 3 moved into lesson 2 package
