+++ src/main/kotlin/lesson_2/lesson_2_2.kt
@@ -1,4 +1,4 @@
-package org.example
+package org.example.lesson_2
 
 fun main() {
